requirements.txt
cupy==11.3.0
cupy_cuda113==9.2.0
h5py==3.1.0
imageio==2.9.0
matplotlib==3.3.4
numpy==1.19.5
opencv_python==4.5.3.56
PyYAML==6.0
scikit_image==0.17.2
scikit_learn==1.1.3
scipy==1.4.1
sigpy==0.1.23
skimage==0.0
torch==1.10.2+cu113
torchsummary==1.5.1
torchvision==0.11.3+cu113
tqdm==4.61.1